Fix bug

+++ graphx/src/main/scala/org/apache/spark/graphx/impl/EdgeTripletIterator.scala
@@ -48,11 +48,11 @@ class EdgeTripletIterator[VD: ClassTag, ED: ClassTag](
   override def next() = {
     triplet.srcId = edgePartition.srcIds(pos)
     // assert(vmap.containsKey(e.src.id))
-    val fakeSrcInd = (triplet.srcId * 1125899906842597L) % vertexArray.length
+    val fakeSrcInd = ((triplet.srcId * 1125899906842597L) % vertexArray.length).toInt
     triplet.srcAttr = vertexArray(fakeSrcInd)
     triplet.dstId = edgePartition.dstIds(pos)
     // assert(vmap.containsKey(e.dst.id))
-    val fakeDstInd = (triplet.dstId * 1125899906842597L) % vertexArray.length
+    val fakeDstInd = ((triplet.dstId * 1125899906842597L) % vertexArray.length).toInt
     triplet.dstAttr = vertexArray(fakeDstInd)
     triplet.attr = edgePartition.data(pos)
     pos += 1
